rename twoLocals to local; make all local based on dev; delete moonbasealpha folder
